+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [r11.2.2] - 2020-01-21
-
 ### Add
 - Background color for empty image in GPS List Store.
 ### Fixed
 - Product has 1 attribute group will auto detect remaining inventory.
+- Fanpage option in section list in Account. (patch: 2020-01-22)
 
 ## [r11.2.1] - 2020-01-20
 ### Added